Turn PayloadMapper and MessageMapper into interfaces
This change turns `PayloadMapper` and `MessageMapper` into interfaces,
and introduces default implementations annotated with `@DefaultBean`.

This allows users to optionally override the logic to serialize and
deserialize payloads by providing an alternate CDI bean. This can
be useful e.g. to inject a different `ObjectMapper`, or to use views
when serializing and deserializing. It also allows users to use
`@Decorator` beans if desired.